+++ components/Signup/Form.tsx
@@ -1,68 +1,15 @@
-import { AccountBookOutlined, LockOutlined, MailOutlined } from "@ant-design/icons";
-import { Form, Row, Col, Divider, Input } from 'antd';
+import React from 'react';
+
+import { Row, Col, Divider } from 'antd';
 
 import { signupUser } from '../../lib/users';
+import { formItems } from '../../lib/constants';
 
-import LoginSection from './Login';
+import FormFields from '../Form/Fields';
+import FormComponent from '../Form/Form';
+
 import Footer from './Footer';
-
-const formItems = [
-    {
-        name: "username",
-        type: "text",
-        rules: [
-            {
-                type: "string",
-                message: "Username is required",
-                required: true,
-            },
-        ],
-        placeholder: "Jane Doe",
-        icon: <AccountBookOutlined />,
-    },
-    {
-        name: "email",
-        type: "text",
-        rules: [
-            {
-                type: "email",
-                message: "Email is required",
-                required: true,
-            },
-        ],
-        placeholder: "someone@example.com",
-        icon: <MailOutlined />,
-    },
-    {
-        name: "password",
-        type: "text",
-        rules: [
-            {
-                type: "string",
-                message: "Password is required",
-                required: true,
-            },
-        ],
-        placeholder: "*********",
-        icon: <LockOutlined />,
-    },
-];
-
-const FormFields = (props) => {
-    const { formItems } = props;
-    return formItems.map(entry => {
-        const { name, rules, placeholder, icon, type } = entry;
-
-        return (
-            <Form.Item
-                name={name}
-                rules={rules}
-            >
-                <Input type={type} placeholder={placeholder} prefix={icon} />
-            </Form.Item>
-        )
-    })
-};
+import LoginSection from './Login';
 
 const FormContainer = () => {
     const formParentStyle = {
@@ -72,8 +19,6 @@ const FormContainer = () => {
         overflow: "hidden",
     };
 
-    const [FormInstance] = Form.useForm();
-
     const signUp = async (details) => {
         await signupUser();
     };
@@ -82,6 +27,20 @@ const FormContainer = () => {
         console.log(error);
     };
 
+    const onFinish = async (values) => {
+        const res = await signUp(values);
+        console.log('res==>', res)
+        // if (errors) {
+        //     return false;
+        // }
+        // if (error?.field) {
+        //     return setAuthError(error?.message);
+        // }
+        // // on successfull account request, inform user to check their email
+        // if (message) {
+        //     return router.push("/instructions?message=Verify your email");
+        // }
+    };
 
     return (
         <Col sm={{ span: 22 }} md={{ span: 18 }} lg={{ span: 14 }}>
@@ -102,31 +61,17 @@ const FormContainer = () => {
                                         {authError && <Alert type="error" showIcon message={authError} />}
                                         {requestMessage && <Alert type="success" showIcon message={requestMessage} />} */}
                         </section>
-                        <Form
-                            form={FormInstance}
+                        <FormComponent
+                            name="signupForm"
                             layout="vertical"
-                            name="signupform"
-                            initialValues={{ remember: true }}
-                            onFinishFailed={onFinishFailed}
-                            onFinish={async values => {
-                                const res = await signUp(values);
-                                console.log('res==>', res)
-                                // if (errors) {
-                                //     return false;
-                                // }
-                                // if (error?.field) {
-                                //     return setAuthError(error?.message);
-                                // }
-                                // // on successfull account request, inform user to check their email
-                                // if (message) {
-                                //     return router.push("/instructions?message=Verify your email");
-                                // }
-                            }}
                             autoComplete="off"
+                            onFinish={onFinish}
+                            onFinishFailed={onFinishFailed}
+                            initialValues={{ remember: true }}
                         >
                             <FormFields formItems={formItems} />
                             <Footer text="Have an account?" cta="Login" cta2="Sign Up" ctaHref="/signin" />
-                        </Form>
+                        </FormComponent>
                     </section>
                 </Col>
             </Row>
